fix: lighting texture and normal front/back face procedural cube
Typo on lighting shaders made prevent texture to be processed.
As we are working on -Z for the forward vector on OpenGL, I set:
x = 0.0, y = 0.0, z = -1.0
For the normal vector, but to solve the problem of direction,
I already change the forward base vector to +Z, so normal of the back
and front faces needed to be inverted.